Make toast params optional and move it under helpers.
