+++ src/domain/auth/authenticate.ts
@@ -1,5 +1,7 @@
 import * as Sentry from '@sentry/browser';
 
+import toastNotification from '../../common/helpers/toastNotification/toastNotification';
+
 import userManager from './userManager';
 import store from '../../redux/store';
 import { apiError } from './redux';
@@ -7,6 +9,7 @@ import { apiError } from './redux';
 function authenticate(): void {
   userManager.signinRedirect().catch(error => {
     if (error.message !== 'Network Error') {
+      toastNotification();
       Sentry.captureException(error);
     }
     store.dispatch(apiError(error.toString()));
