+++ src/domain/landingPage/LandingPage.tsx
@@ -3,12 +3,12 @@ import React from 'react';
 import MembershipInformationPage from '../membership/information/MembershipInformationPage';
 import useIsMembershipPending from '../membership/useIsMembershipPending';
 import SentYouthProfilePage from '../youthProfile/sent/SentYouthProfilePage';
-import toastNotification from '../../common/components/notification/toastNotification';
+import toastNotification from '../../common/helpers/toastNotification/toastNotification';
 
 function LandingPage() {
   const [isMembershipPending, loading] = useIsMembershipPending({
     onError: () => {
-      toastNotification({});
+      toastNotification();
     },
   });
